// src/auth/enums/permission-type.enum.ts
export enum PermissionType {
  ViewIncome = 'view_income',
  ViewSales = 'view_sales',
  CreateSales = 'create_sales',
  UpdateSales = 'update_sales',
  DeleteSales = 'delete_sales',
  ViewBusiness = 'view_business',
  CreateBusiness = 'create_business',
  UpdateBusiness = 'update_business',
  DeleteBusiness = 'delete_business',
}